Setup workflow with private tests repo (#1616)
* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Extend CI For private tests

* Add npm script test:fetch-private

* Adjust eslintingore

* Adjust test.yml GH workflow

* Add performance.yml GH workflow

* Setup codecov.yml

* Remove Makefile

* Bring back removed npm scripts

* Update test/README.md

* Add setup files for jest and karma

* Move codecov.yml to the repository root

* Fix typo in eslintignore file

---------
