+++ .eslintignore
@@ -15,6 +15,7 @@ test/_setupFiles/*.js
 
 # Auto-generated directories
 commonjs
+coverage
 dist
 doc
 es
