+++ .github/workflows/build-docs.yml
@@ -23,17 +23,15 @@ jobs:
     name: build-docs
     runs-on: ${{ matrix.os }}
     steps:
-      - uses: actions/checkout@722adc63f1aa60a57ec37892e133b1d319cae598 # https://github.com/actions/checkout/releases/tag/v2.0.0
-
       - name: Setup Node.js ${{ matrix.node-version }}
         uses: actions/setup-node@56899e050abffc08c2b3b61f3ec6a79a9dc3223d # https://github.com/actions/setup-node/releases/tag/v1.4.4
         with:
           node-version: ${{ matrix.node-version }}
 
+      - uses: actions/checkout@722adc63f1aa60a57ec37892e133b1d319cae598 # https://github.com/actions/checkout/releases/tag/v2.0.0
+
       - name: Install dependencies
-        run: |
-          npm ci
+        run: npm ci
 
       - name: Build docs
-        run: |
-          npm run docs:build
+        run: npm run docs:build
