+++ codecov.yml
@@ -1,0 +1,14 @@
+codecov:
+  require_ci_to_pass: yes
+
+coverage:
+  range: 95..100
+  round: down
+  precision: 2
+
+comment:
+  layout: "reach, diff, flags, files"
+  behavior: new
+  require_changes: false
+  require_base: yes
+  require_head: yes
