+++ package.json
@@ -76,14 +76,26 @@
     "verify:publish-package": "npm pack | node script/check-publish-package.js",
     "verify:typings": "tsc --noEmit",
     "test": "npm-run-all lint test:jest test:browser",
+    "test:fetch-private": "bash test/fetch-tests.sh",
     "test:jest": "cross-env NODE_ICU_DATA=node_modules/full-icu jest",
+    "test:watch": "npm run test:jest -- --watch",
+    "test:tdd": "npm run test:jest -- --watch function-value",
+    "test:coverage": "npm run test:jest -- --coverage",
+    "test:logMemory": "npm run test:jest -- --runInBand --logHeapUsage",
+    "test:performance": "npm run benchmark:basic && npm run benchmark:cruds",
+    "test:compatibility": "bash test/compatibility/test-compatibility.sh",
     "test:ci": "cross-env NODE_ICU_DATA=node_modules/full-icu node --expose-gc ./node_modules/jest/bin/jest --forceExit",
     "test:browser": "cross-env-shell BABEL_ENV=dist env-cmd -f ht.config.js karma start",
+    "test:browser.debug": "cross-env-shell BABEL_ENV=dist NODE_ENV=debug env-cmd -f ht.config.js karma start",
     "typedoc:build-api": "cross-env NODE_OPTIONS=--openssl-legacy-provider typedoc --options .typedoc.md.ts",
+    "benchmark:basic": "npm run tsnode test/performance/run-basic-benchmark.ts",
+    "benchmark:cruds": "npm run tsnode test/performance/run-cruds-benchmark.ts",
+    "benchmark:write-to-file": "npm run tsnode test/performance/write-to-file.ts",
+    "benchmark:compare-benchmarks": "npm run tsnode test/performance/compare-benchmarks.ts",
     "lint": "eslint . --ext .js,.ts",
     "lint:fix": "eslint . --ext .js,.ts --fix",
     "audit": "npm audit --omit=dev",
-    "clean": "rimraf commonjs/ dist/ es/ languages/ lib/ typings/ test-jasmine/",
+    "clean": "rimraf coverage/ commonjs/ dist/ es/ languages/ lib/ typings/ test-jasmine/",
     "compile": "tsc",
     "check:licenses": "license-checker --production --excludePackages=\"hyperformula@3.1.1\" --onlyAllow=\"MIT; Apache-2.0; BSD-3-Clause; BSD-2-Clause; ISC; BSD; Unlicense\"",
     "tsnode": "ts-node --transpile-only -O {\\\"module\\\":\\\"commonjs\\\"}"
